src/pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = "Template for new python projects"
authors = ["Andrey Volkov <andvolkov.dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {version = "^0.104.1"}
asyncpg = "^0.29.0"
alembic = "^1.13.0"
jinja2 = "^3.1.2"
greenlet = "^3.0.1"
uvicorn = "^0.25.0"
gunicorn = "^21.2.0"
sqlmodel = "^0.0.14"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.1.0"
pycryptodome = "^3.19.1"
sqlalchemy = "^2.0.24"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^2.6.1"
ruff = "^0.1.9"
isort = "^5.12.3"
mypy = "^1.7.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
# Enable pycodestyle (`E`) and Pyflakes (`F`) codes by default.
select = ["E", "F", "Q"]
ignore = []

# Allow autofix for all enabled rules (when `--fix`) is provided.
fixable = ["A", "B", "C", "D", "E", "F", "Q"]
unfixable = []

# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 79

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# Assume Python 3.12.
target-version = "py312"

[tool.ruff.mccabe]
max-complexity = 10
